requirements.txt
pyyaml>=6.0.1
requests>=2.31.0
Flask>=2.3.2